Same record structure for primary tree.
The primary and staging trees now share the same record structure. This
adds an extra, unnecessary property to the primary tree. Actually, it is
a property of each record, but it is known to be zero. Is is there, but
we don't need to store it. Thus, storing it doesn't feel so wrong,
because it is there after all.

In many cases, it would compress well if we start gzipping our leaves.

Closes #47.