Left most iterator with no range.
Implement a left most iterator. When no key is provided for the
iterator, go to the left most key. When no range is provided, see that
the iterator will procede to the end of the collection.

Note that the iterator returns nulls to indicate an end of the
iteration, while the absence of a record when invoking `.get` raises and
error.

Closes #60.
Closes #59.