Extract MVCC iterator creation to Locket object.
We're going to want to create the MVCC iterator for `approximateSize`
and sum the underlying record sizes. We'll want the exact same sort of
iterator that the user will create using `Locket.iterator` so we'll move
the MVCC iterator creation into `Locket`. The `Iterator` now asks
`Locket` to build an iterator.

Closes #102.