Implement simple incoming log, read performance.
This update ingests all the upstream improvements to Cadence, Strata,
and the MVCC support libraries. Performance is still much slower than
LevelDB for our once benchmark, but it is not as deathly slow as it was.
There will be further performance improvements, but they should be based
on more real world use cases, not just the one load and read benchmark.

Locket performance is improved by

 - reducing asynchronous operations in Strata in MVCC libraries,
 - drastically improved performance of the Cadence library,
 - new binary file format,
 - simplified batch operation intake.

Locket performance can be further improved by reducing the decorator
patterns used to create iterators, maybe. Performing the data validation
of batch operations of the `abstract-leveldown` library inside Locket
while inserting, reducing a loop over the data. That's what's coming to
me at the moment.

Currently, merge is unimplemented. It needs to be re-implemented.

Closes #197.
Closes #196.
Closes #138.
Closes #119.
Closes #23.
Closes #19.