Upgrade to trampoline based Amalgamate.
Created a trampoline to use with the MVCC iterator collection. An async
call will only be performed if necessary, if an asynchronous file read
in required. If the results in the the cache, a descent of the b-tree is
synchronous. Reduces trips to the async event queue. Allows us to use
the decorator pattern for iterators without nesting `async` calls. Any
`async` call is exported through the trampoline. If there are `async`
calls in the trampoline they are awaited `await`ed, if not they are
ignored.

Closes #247.