Convert to Cadence.
Since we're implementing an error-first callback interface, let's do so.
I've added promise support to Cadence, and error-first callback support
to Reciprocate. Now there ought to be no unnecessary escapes to the
Node.js event loop or the promise event loop. Anything that can be run
synchrnously will be run synchronously using the trampoline in Cadence
or the trampoline in Reciprocate.

Closes #249.