Upgrade `satiate` to `0.0.1`.
Direction is now a property of iterators. It is one of `mvcc.FORWARD`,
`mvcc.REVERSE` or `mvcc.MAP` defined in the `mvcc` constants module.
Satiate will simply inherit the direction of the decorated iterator. It
is not relevant yet because we do not decorate our Satiate iterator and
the Satiate iterator works regardless of direction.

https://github.com/bigeasy/satiate/releases/tag/v0.0.1

Closes #252.